refactor(core): add operation concept(2)
