+++ core/flamingock-core/src/main/java/io/flamingock/internal/core/operation/apply/ApplyResult.java
@@ -1,0 +1,6 @@
+package io.flamingock.internal.core.operation.apply;
+
+import io.flamingock.internal.core.operation.AbstractOperationResult;
+
+public class ApplyResult extends AbstractOperationResult {
+}
