--- core/flamingock-core/src/main/java/io/flamingock/internal/core/runner/apply/ApplyResult.java
@@ -1,6 +1,0 @@
-package io.flamingock.internal.core.runner.apply;
-
-import io.flamingock.internal.core.runner.AbstractOperationResult;
-
-public class ApplyResult extends AbstractOperationResult {
-}
