+++ core/target-systems/sql-target-system/src/test/java/io/flamingock/targetsystem/sql/SqlTargetSystemTest.java
@@ -33,7 +33,7 @@ import io.flamingock.internal.common.cloud.vo.TargetSystemAuditMarkType;
 import io.flamingock.internal.core.builder.FlamingockFactory;
 import io.flamingock.internal.core.builder.CloudChangeRunnerBuilder;
 import io.flamingock.internal.common.core.util.Deserializer;
-import io.flamingock.internal.core.runner.OperationException;
+import io.flamingock.internal.core.operation.OperationException;
 import io.flamingock.internal.core.builder.runner.Runner;
 import org.junit.jupiter.api.*;
 import org.mockito.MockedStatic;
